fix video size is in px but set as dp
