+++ feature/player/src/main/java/dev/anilbeesetti/nextplayer/feature/player/ui/VerticalProgressView.kt
@@ -1,0 +1,95 @@
+package dev.anilbeesetti.nextplayer.feature.player.ui
+
+import androidx.annotation.IntRange
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.text.BasicText
+import androidx.compose.foundation.text.TextAutoSize
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import dev.anilbeesetti.nextplayer.core.ui.R
+import dev.anilbeesetti.nextplayer.core.ui.theme.NextPlayerTheme
+
+@Composable
+fun VerticalProgressView(
+    modifier: Modifier = Modifier,
+    width: Dp = 32.dp,
+    icon: Painter,
+    @IntRange(from = 0, to = 100) value: Int,
+) {
+    Column(
+        modifier = modifier
+            .heightIn(max = 250.dp)
+            .clip(CircleShape)
+            .background(MaterialTheme.colorScheme.background)
+            .padding(8.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(4.dp),
+    ) {
+        Box(
+            modifier = Modifier.size(width),
+            contentAlignment = Alignment.Center
+        ) {
+            BasicText(
+                text = value.toString(),
+                style = MaterialTheme.typography.labelLarge.copy(
+                    color = MaterialTheme.colorScheme.onBackground
+                ),
+                autoSize = TextAutoSize.StepBased(maxFontSize = MaterialTheme.typography.labelLarge.fontSize),
+            )
+        }
+        Box(
+            modifier = Modifier
+                .weight(1f)
+                .width(width)
+                .clip(MaterialTheme.shapes.medium)
+                .background(MaterialTheme.colorScheme.surfaceVariant),
+            contentAlignment = Alignment.BottomCenter,
+        ) {
+            Box(
+                modifier = Modifier
+                    .width(width)
+                    .fillMaxHeight(value / 100f)
+                    .background(MaterialTheme.colorScheme.primary),
+            )
+        }
+        Box(
+            modifier = Modifier.size(width),
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                painter = icon,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.onBackground,
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun VerticalProgressPreview() {
+    NextPlayerTheme {
+        VerticalProgressView(
+            value = 50,
+            icon = painterResource(R.drawable.ic_volume)
+        )
+    }
+}
