+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 android-compileSdk = "36"
 android-targetSdk = "36"
-android-minSdk = "21"
+android-minSdk = "23"
 android-jvm = "17"
 aboutlibraries = "11.2.3"
 accompanist = "0.37.3"
@@ -17,7 +17,7 @@ androidxDataStore = "1.1.7"
 androidxEspresso = "3.6.1"
 androidxHiltNavigationCompose = "1.2.0"
 androidxLifecycle = "2.9.1"
-androidxMedia3 = "1.8.0"
+androidxMedia3 = "1.9.0"
 androidxNavigation = "2.9.1"
 androidxTestCore = "1.5.0"
 androidxTestExt = "1.2.1"
